Refactoring middleware functions to their own tests
