+++ test/unit-tests/middleware/validateJsonSchema4.0.js
@@ -1,0 +1,129 @@
+const express = require('express')
+const app = express()
+const chai = require('chai')
+const expect = chai.expect
+chai.use(require('chai-http'))
+
+// Body Parser Middleware
+app.use(express.json()) // Allows us to handle raw JSON data
+app.use(express.urlencoded({ extended: false })) // Allows us to handle url encoded data
+const middleware = require('../../../src/middleware/middleware')
+app.use(middleware.createCtxAndReqUUID)
+
+const mwFixtures = require('./mockObjects.middleware')
+const cveId4 = 'CVE-2017-5832'
+const cvePass4 = require('../../schemas/4.0/' + cveId4 + '_public.json')
+const cveFail4 = require('../../schemas/4.0/' + cveId4 + '_fail.json')
+const cveRejectFail4 = require('../../schemas/4.0/' + cveId4 + '_reject_fail.json')
+const cveReservedFail4 = require('../../schemas/4.0/' + cveId4 + '_reserved_fail.json')
+const cvePublicFail4 = require('../../schemas/4.0/' + cveId4 + '_public_fail.json')
+
+app.route('/api/test/mw/schema4')
+  .post(middleware.validateCveJsonSchema, (req, res) => {
+    return res.status(200).json({ message: 'Success! You have reached the target endpoint.' })
+  })
+
+describe('Test the JSON schema 4.0 validation middleware', () => {
+  context('Validate JSON schema 4.0', () => {
+    it('Json validator pass', (done) => {
+      chai.request(app)
+        .post('/api/test/mw/schema4')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cvePass4)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(200)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('Success! You have reached the target endpoint.')
+          done()
+        })
+    })
+
+    it('Json validator fails because the STATE is invalid', (done) => {
+      chai.request(app)
+        .post('/api/test/mw/schema4')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cveFail4)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('CVE JSON schema validation FAILED.')
+          expect(res.body.details).to.have.property('errors').and.to.be.an('array')
+          expect(res.body.details.errors[0]).to.have.string('CVE_data_meta.STATE is not one of enum values')
+          done()
+        })
+    })
+
+    it('Json validator fails in REJECT state', (done) => {
+      chai.request(app)
+        .post('/api/test/mw/schema4')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cveRejectFail4)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.property('status', 400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('CVE JSON schema validation FAILED.')
+          expect(res.body.details).to.have.property('errors').to.be.an('array')
+          expect(res.body.details.errors[0]).to.have.string('instance.description.description_data[0] requires property "lang"')
+          done()
+        })
+    })
+
+    it('Json validator fails in RESERVED state', (done) => {
+      chai.request(app)
+        .post('/api/test/mw/schema4')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cveReservedFail4)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.property('status', 400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('CVE JSON schema validation FAILED.')
+          expect(res.body.details).to.have.property('errors').to.be.an('array')
+          expect(res.body.details.errors[0]).to.have.string('instance.description.description_data does not meet minimum length of 1')
+          done()
+        })
+    })
+
+    it('Json validator fails in PUBLIC state', (done) => {
+      chai.request(app)
+        .post('/api/test/mw/schema4')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cvePublicFail4)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('CVE JSON schema validation FAILED.')
+          expect(res.body.details).to.have.property('errors').and.to.be.an('array')
+          expect(res.body.details.errors[0]).to.have.string('instance.affects.vendor.vendor_data[0].product.product_data[0].version.version_data[0] requires property "version_value"')
+          expect(res.body.details.errors[1]).to.have.string('instance.affects.vendor.vendor_data[0].product.product_data[1] requires property "version"')
+          expect(res.body.details.errors[2]).to.have.string('instance.problemtype.problemtype_data[0].description[0] requires property "lang"')
+          expect(res.body.details.errors[3]).to.have.string('instance.description.description_data is not of a type(s) array')
+          done()
+        })
+    })
+  })
+})
