+++ test/unit-tests/middleware/validateJsonSchema5.0.js
@@ -1,0 +1,131 @@
+const express = require('express')
+const app = express()
+const chai = require('chai')
+const expect = chai.expect
+chai.use(require('chai-http'))
+
+// Body Parser Middleware
+app.use(express.json()) // Allows us to handle raw JSON data
+app.use(express.urlencoded({ extended: false })) // Allows us to handle url encoded data
+const middleware = require('../../../src/middleware/middleware')
+app.use(middleware.createCtxAndReqUUID)
+
+const mwFixtures = require('./mockObjects.middleware')
+const cveId5 = 'CVE-2017-4024'
+const cvePass5 = require('../../schemas/5.0/' + cveId5 + '_public.json')
+const cveFail5 = require('../../schemas/5.0/' + cveId5 + '_fail.json')
+const cveRejectFail5 = require('../../schemas/5.0/' + cveId5 + '_reject_fail.json')
+const cveReservedFail5 = require('../../schemas/5.0/' + cveId5 + '_reserved_fail.json')
+const cvePublicFail5 = require('../../schemas/5.0/' + cveId5 + '_public_fail.json')
+
+app.route('/api/test/mw/schema5')
+  .post(middleware.validateCveJsonSchema, (req, res) => {
+    return res.status(200).json({ message: 'Success! You have reached the target endpoint.' })
+  })
+
+describe('Test the JSON schema 5.0 validation middleware', () => {
+  context('Validate JSON schema 5.0', () => {
+    it('Json validator pass', (done) => {
+      chai.request(app)
+        .post('/api/test/mw/schema5')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cvePass5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(200)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('Success! You have reached the target endpoint.')
+          done()
+        })
+    })
+
+    it('Json validator fails because the STATE is invalid', (done) => {
+      chai.request(app)
+        .post('/api/test/mw/schema5')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cveFail5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('CVE JSON schema validation FAILED.')
+          expect(res.body.details).to.have.property('errors').and.to.be.an('array')
+          expect(res.body.details.errors[0]).to.have.string('CVE_data_meta.STATE is not one of enum values')
+          done()
+        })
+    })
+
+    it('Json validator fails in REJECT state', (done) => {
+      chai.request(app)
+        .post('/api/test/mw/schema5')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cveRejectFail5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.property('status', 400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('CVE JSON schema validation FAILED.')
+          expect(res.body.details).to.have.property('errors').to.be.an('array')
+          expect(res.body.details.errors[0]).to.have.string('instance.descriptions[0] requires property "lang"')
+          done()
+        })
+    })
+
+    it('Json validator fails in RESERVED state', (done) => {
+      chai.request(app)
+        .post('/api/test/mw/schema5')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cveReservedFail5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.property('status', 400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('CVE JSON schema validation FAILED.')
+          expect(res.body.details).to.have.property('errors').to.be.an('array')
+          expect(res.body.details.errors[0]).to.have.string('value does not meet minimum length of 1')
+          expect(res.body.details.errors[1]).to.have.string('instance.descriptions[0] requires property "lang"')
+          done()
+        })
+    })
+
+    it('Json validator fails in PUBLIC state', (done) => {
+      chai.request(app)
+        .post('/api/test/mw/schema5')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cvePublicFail5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('CVE JSON schema validation FAILED.')
+          expect(res.body.details).to.have.property('errors').and.to.be.an('array')
+          expect(res.body.details.errors[0]).to.have.string('CNA.provider_data_meta.ID is not of a type(s) string')
+          expect(res.body.details.errors[1]).to.have.string('CNA.affected.vendors[0].products[0].versions[0].version_affected is not one of enum values')
+          expect(res.body.details.errors[2]).to.have.string('CNA.affected.affects_SWID does not meet minimum length of 1')
+          expect(res.body.details.errors[3]).to.have.string('CNA.problemtypes contains duplicate item')
+          expect(res.body.details.errors[4]).to.have.string('additionalProperty "additional" exists in instance when not allowed')
+          done()
+        })
+    })
+  })
+})
