+++ test/unit-tests/middleware/validateUserTest.js
@@ -1,0 +1,295 @@
+const express = require('express')
+const app = express()
+const chai = require('chai')
+const expect = chai.expect
+chai.use(require('chai-http'))
+
+// Body Parser Middleware
+app.use(express.json()) // Allows us to handle raw JSON data
+app.use(express.urlencoded({ extended: false })) // Allows us to handle url encoded data
+const middleware = require('../../../src/middleware/middleware')
+app.use(middleware.createCtxAndReqUUID)
+
+const CONSTANTS = require('../../../src/constants')
+const errors = require('../../../src/middleware/error')
+const error = new errors.MiddlewareError()
+const repos = require('../../../test-utils/repositories')
+
+const mwFixtures = require('./mockObjects.middleware')
+const cveId5 = 'CVE-2017-4024'
+const cvePass5 = require('../../schemas/5.0/' + cveId5 + '_public.json')
+
+describe('Testing the user validation middleware', () => {
+  context('Validate user', () => {
+    it('User is successfully validated', (done) => {
+      app.route('/validate-user-successful')
+        .post((req, res, next) => {
+          const factory = {
+            getOrgRepository: () => { return new repos.OrgValidateUserSuccess() },
+            getUserRepository: () => { return new repos.UserValidateUserSuccess() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, middleware.validateUser, (req, res) => {
+          return res.status(200).json({ message: 'Success! You have reached the target endpoint.' })
+        })
+
+      chai.request(app)
+        .post('/validate-user-successful')
+        .set(mwFixtures.secretariatHeaders)
+        .send(cvePass5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(200)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('message').and.to.be.a('string')
+          expect(res.body.message).to.equal('Success! You have reached the target endpoint.')
+          done()
+        })
+    })
+
+    it('Org does not exist', (done) => {
+      app.route('/validate-user-org-doesnt-exist')
+        .post((req, res, next) => {
+          const factory = {
+            getOrgRepository: () => { return new repos.NullOrgRepo() },
+            getUserRepository: () => { return new repos.NullUserRepo() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, middleware.validateUser, (req, res) => {
+          return res.status(200).json({ message: 'Success! You have reached the target endpoint.' })
+        })
+
+      const testHeaders = Object.assign({}, mwFixtures.secretariatHeaders)
+      testHeaders[CONSTANTS.AUTH_HEADERS.ORG] = 'jpmorgan'
+
+      chai.request(app)
+        .post('/validate-user-org-doesnt-exist')
+        .set(testHeaders)
+        .send(cvePass5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(401)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          const errObj = error.unauthorized()
+          expect(res.body.error).to.equal(errObj.error)
+          expect(res.body.message).to.equal(errObj.message)
+          done()
+        })
+    })
+
+    it('User is not found', (done) => {
+      app.route('/validate-user-user-not-found')
+        .post((req, res, next) => {
+          const factory = {
+            getOrgRepository: () => { return new repos.NullOrgRepo() },
+            getUserRepository: () => { return new repos.NullUserRepo() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, middleware.validateUser, (req, res) => {
+          return res.status(200).json({ message: 'Success! You have reached the target endpoint.' })
+        })
+
+      const testHeaders = Object.assign({}, mwFixtures.secretariatHeaders)
+      testHeaders[CONSTANTS.AUTH_HEADERS.USER] = 'morgan'
+
+      chai.request(app)
+        .post('/validate-user-user-not-found')
+        .set(testHeaders)
+        .send(cvePass5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(401)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          const errObj = error.unauthorized()
+          expect(res.body.error).to.equal(errObj.error)
+          expect(res.body.message).to.equal(errObj.message)
+          done()
+        })
+    })
+
+    it('Secret apikey is incorrect', (done) => {
+      app.route('/validate-user-key-incorrect')
+        .post((req, res, next) => {
+          const factory = {
+            getOrgRepository: () => { return new repos.OrgValidateUserSuccess() },
+            getUserRepository: () => { return new repos.UserValidateUserSuccess() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, middleware.validateUser, (req, res) => {
+          return res.status(200).json({ message: 'Success! You have reached the target endpoint.' })
+        })
+
+      const testHeaders = Object.assign({}, mwFixtures.secretariatHeaders)
+      testHeaders[CONSTANTS.AUTH_HEADERS.KEY] = 'wrong secret'
+
+      chai.request(app)
+        .post('/validate-user-key-incorrect')
+        .set(testHeaders)
+        .send(cvePass5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(401)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          const errObj = error.unauthorized()
+          expect(res.body.error).to.equal(errObj.error)
+          expect(res.body.message).to.equal(errObj.message)
+          done()
+        })
+    })
+
+    it('User is deactivated', (done) => {
+      app.route('/validate-user-deactivated')
+        .post((req, res, next) => {
+          const factory = {
+            getOrgRepository: () => { return new repos.OrgValidateUserSuccess() },
+            getUserRepository: () => { return new repos.UserValidateUserDeactivated() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, middleware.validateUser, (req, res) => {
+          return res.status(200).json({ message: 'Success! You have reached the target endpoint.' })
+        })
+
+      const deactivatedHeaders = {}
+      deactivatedHeaders[CONSTANTS.AUTH_HEADERS.ORG] = 'mitre'
+      deactivatedHeaders[CONSTANTS.AUTH_HEADERS.KEY] = 'S96E4QT-SMT4YE3-KX03X6K-4615CED'
+      deactivatedHeaders[CONSTANTS.AUTH_HEADERS.USER] = 'flast'
+      deactivatedHeaders['content-type'] = 'application/json'
+
+      chai.request(app)
+        .post('/validate-user-deactivated')
+        .set(deactivatedHeaders)
+        .send()
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(401)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          const errObj = error.unauthorized()
+          expect(res.body.error).to.equal(errObj.error)
+          expect(res.body.message).to.equal(errObj.message)
+          done()
+        })
+    })
+
+    it('Org is undefined', (done) => {
+      app.route('/validate-user-org-undefined')
+        .post((req, res, next) => {
+          const factory = {
+            getOrgRepository: () => { return new repos.NullOrgRepo() },
+            getUserRepository: () => { return new repos.NullUserRepo() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, middleware.validateUser, (req, res) => {
+          return res.status(200).json({ message: 'Success! You have reached the target endpoint.' })
+        })
+
+      const testHeaders = Object.assign({}, mwFixtures.secretariatHeaders)
+      delete testHeaders[CONSTANTS.AUTH_HEADERS.ORG]
+
+      chai.request(app)
+        .post('/validate-user-org-undefined')
+        .set(testHeaders)
+        .send(cvePass5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          const errObj = error.badRequest(CONSTANTS.AUTH_HEADERS.ORG)
+          expect(res.body.error).to.equal(errObj.error)
+          expect(res.body.message).to.equal(errObj.message)
+          done()
+        })
+    })
+
+    it('Requester is undefined', (done) => {
+      app.route('/validate-user-submitter-undefined')
+        .post((req, res, next) => {
+          const factory = {
+            getOrgRepository: () => { return new repos.NullOrgRepo() },
+            getUserRepository: () => { return new repos.NullUserRepo() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, middleware.validateUser, (req, res) => {
+          return res.status(200).json({ message: 'Success! You have reached the target endpoint.' })
+        })
+
+      const testHeaders = Object.assign({}, mwFixtures.secretariatHeaders)
+      delete testHeaders[CONSTANTS.AUTH_HEADERS.USER]
+
+      chai.request(app)
+        .post('/validate-user-submitter-undefined')
+        .set(testHeaders)
+        .send(cvePass5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          const errObj = error.badRequest(CONSTANTS.AUTH_HEADERS.USER)
+          expect(res.body.error).to.equal(errObj.error)
+          expect(res.body.message).to.equal(errObj.message)
+          done()
+        })
+    })
+
+    it('Secret apikey is undefined', (done) => {
+      app.route('/validate-user-secret-undefined')
+        .post((req, res, next) => {
+          const factory = {
+            getOrgRepository: () => { return new repos.NullOrgRepo() },
+            getUserRepository: () => { return new repos.NullUserRepo() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, middleware.validateUser, (req, res) => {
+          return res.status(200).json({ message: 'Success! You have reached the target endpoint.' })
+        })
+
+      const testHeaders = Object.assign({}, mwFixtures.secretariatHeaders)
+      delete testHeaders[CONSTANTS.AUTH_HEADERS.KEY]
+
+      chai.request(app)
+        .post('/validate-user-secret-undefined')
+        .set(testHeaders)
+        .send(cvePass5)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          expect(res).to.have.status(400)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          const errObj = error.badRequest(CONSTANTS.AUTH_HEADERS.KEY)
+          expect(res.body.error).to.equal(errObj.error)
+          expect(res.body.message).to.equal(errObj.message)
+          done()
+        })
+    })
+  })
+})
